Introduced request package (corresponding to go-ns/common with a bit of refactoring), to prevent cyclical dependency with log.go, update log.go dependency. Made server middleware list and order not exported, to force callers to create their own middelware chain, which will always be executed after requestID and log middleware.
