+++ go.mod
@@ -3,7 +3,7 @@ module github.com/ONSdigital/dp-net
 go 1.13
 
 require (
-	github.com/ONSdigital/log.go v1.0.0
+	github.com/ONSdigital/log.go v1.0.1-0.20200805145532-1f25087a0744
 	github.com/facebookgo/freeport v0.0.0-20150612182905-d4adf43b75b9
 	github.com/justinas/alice v1.2.0
 	github.com/smartystreets/goconvey v1.6.4
