--- http/collection.go
@@ -1,4 +1,0 @@
-package http
-
-const CollectionIDHeaderKey = "Collection-Id"
-const CollectionIDCookieKey = "collection"
