+++ request/collection.go
@@ -1,0 +1,7 @@
+package request
+
+// CollectionID header and cookie keys
+const (
+	CollectionIDHeaderKey = "Collection-Id"
+	CollectionIDCookieKey = "collection"
+)
