+++ request/locale.go
@@ -1,4 +1,4 @@
-package http
+package request
 
 import (
 	"net/http"
